separated AlignmentFile and AlignedSegment testing
